requirements.txt
Pillow
click

